[REFACTOR]: estrutura de pastas de NavBar
