+++ components/NavBar/Menu/index.tsx
@@ -1,11 +1,7 @@
-import { Link } from 'lib/links';
 import { useState } from 'react';
 import { FaBars, FaTimes } from 'react-icons/fa';
+import { MenuProps } from '../intefaces';
 import Links from '../Links';
-
-export interface MenuProps {
-  links: Link[];
-}
 
 export default function Menu({ links }: MenuProps) {
   const [nav, setNav] = useState(false);
